Fix TextAttributeKey registration error
Somehow, I was getting an error saying that the "IDENTIFIER"
TextAttributeKey had already been registered with another fallback
attribute key: "DEFAULT_KEYWORD". I'm not sure how that was happening so
I've just changed this key to "USD.IDENTIFIER" to avoid the name
collision; that seemed to fix/avoid the problem.